refactor: AccAmenityController 반환을 view > data로 변경
