+++ src/main/resources/mappers/AccAmenity.xml
@@ -3,7 +3,7 @@
         PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
         "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
 
-<mapper namespace="org.spring.example.accAmenity.mapper.AccAmenityMapper">
+<mapper namespace="org.spring.example.mapper.AccAmenityMapper">
 
     <select id="findByAccId" parameterType="long" resultType="org.spring.example.accAmenity.domain.AccAmenity">
         SELECT * FROM acc_amenity WHERE acc_id = #{accId}
