+++ src/main/resources/mappers/AmenityCategory.xml
@@ -3,7 +3,7 @@
         PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
         "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
 
-<mapper namespace="org.spring.example.amenityCategory.mapper.AmenityCategoryMapper">
+<mapper namespace="org.spring.example.mapper.AmenityCategoryMapper">
 
     <select id="findAll" resultType="org.spring.example.amenityCategory.domain.AmenityCategory">
         SELECT * FROM amenity_category
