+++ src/main/resources/mappers/AmenityMapper.xml
@@ -3,7 +3,7 @@
         PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
         "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
 
-<mapper namespace="org.spring.example.amenity.mapper.AmenityMapper">
+<mapper namespace="org.spring.example.mapper.AmenityMapper">
 
     <select id="findAll" resultType="org.spring.example.amenity.domain.Amenity">
         SELECT * FROM amenity
